remove ... good bye beacon styles

+++ src/Covid19Radar.Api.Common/Models/SequenceModel.cs
@@ -8,6 +8,6 @@ namespace Covid19Radar.Api.Models
     {
         public string id { get; set; }
         public string PartitionKey { get; set; }
-        public int value { get; set; }
+        public ulong value { get; set; }
     }
 }
